slight refactor
